Implement a counter and headers handling.
The counter is a simple count_all with specifiable name, it passes through all
NextDoc queries and counts non-nil returns.

Headers are collected by a simple extension to QueryOp - ProcessHeaders which
receives a simple interface with the Add method that takes a key/value
pair for headers that need to be added.

The header collector is implemented with an obvious map[string]string.

Currently we output headers after the result. debugging.